requirements.txt
fastapi
sys
csv
datetime
fastapi.responses
numpy
uvicorn
io
mediapipe
mediapipe.tasks
mediapipe.tasks.python
pillow
google-cloud-vision
python-multipart
openai
python-dotenv
starlette.middlewares.cors